move tests to runtime-tests
