+++ runtime/integration-tests/src/lib.rs
@@ -16,9 +16,6 @@
 
 mod evm;
 mod liquidity_pools;
-mod loans {
-	mod development;
-}
 mod pools;
 mod rewards;
 mod runtime_apis;
